Version allégée de App.js pour publicodes
